+++ src/main/java/io/supertokens/storage/mysql/queries/GeneralQueries.java
@@ -27,6 +27,7 @@ import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicExceptio
 import io.supertokens.pluginInterface.multitenancy.AppIdentifier;
 import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
 import io.supertokens.storage.mysql.ConnectionPool;
+import io.supertokens.storage.mysql.PreparedStatementValueSetter;
 import io.supertokens.storage.mysql.Start;
 import io.supertokens.storage.mysql.config.Config;
 import io.supertokens.storage.mysql.utils.Utils;
@@ -44,8 +45,7 @@ import java.util.stream.Collectors;
 import static io.supertokens.storage.mysql.PreparedStatementValueSetter.NO_OP_SETTER;
 import static io.supertokens.storage.mysql.ProcessState.PROCESS_STATE.CREATING_NEW_TABLE;
 import static io.supertokens.storage.mysql.ProcessState.getInstance;
-import static io.supertokens.storage.mysql.QueryExecutorTemplate.execute;
-import static io.supertokens.storage.mysql.QueryExecutorTemplate.update;
+import static io.supertokens.storage.mysql.QueryExecutorTemplate.*;
 import static io.supertokens.storage.mysql.queries.EmailPasswordQueries.getQueryToCreatePasswordResetTokenExpiryIndex;
 import static io.supertokens.storage.mysql.queries.EmailPasswordQueries.getQueryToCreatePasswordResetTokensTable;
 import static io.supertokens.storage.mysql.queries.EmailVerificationQueries.*;
@@ -256,6 +256,7 @@ public class GeneralQueries {
 
             // index
             update(con, getQueryToCreateSessionExpiryIndex(start), NO_OP_SETTER);
+            update(con, getQueryToCreateSessionAppIdUserIdIndex(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getTenantConfigsTable())) {
@@ -288,12 +289,18 @@ public class GeneralQueries {
         if (!doesTableExists(start, con, Config.getConfig(start).getEmailPasswordUsersTable())) {
             getInstance(start).addState(CREATING_NEW_TABLE, null);
             update(con, EmailPasswordQueries.getQueryToCreateUsersTable(start), NO_OP_SETTER);
+
+            // index
+            update(con, EmailPasswordQueries.getQueryToCreateEmailPasswordUsersEmailIndex(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getEmailPasswordUserToTenantTable())) {
             getInstance(start).addState(CREATING_NEW_TABLE, null);
             update(con, EmailPasswordQueries.getQueryToCreateEmailPasswordUserToTenantTable(start),
                     NO_OP_SETTER);
+
+            // index
+            update(con, EmailPasswordQueries.getQueryToCreateEmailPasswordUserToTenantEmailIndex(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getPasswordResetTokensTable())) {
@@ -329,6 +336,9 @@ public class GeneralQueries {
         if (!doesTableExists(start, con, Config.getConfig(start).getThirdPartyUserToTenantTable())) {
             getInstance(start).addState(CREATING_NEW_TABLE, null);
             update(con, ThirdPartyQueries.getQueryToCreateThirdPartyUserToTenantTable(start), NO_OP_SETTER);
+
+            // index
+            update(con, ThirdPartyQueries.getQueryToCreateThirdPartyUserToTenantThirdPartyUserIdIndex(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getJWTSigningKeysTable())) {
@@ -339,12 +349,20 @@ public class GeneralQueries {
         if (!doesTableExists(start, con, Config.getConfig(start).getPasswordlessUsersTable())) {
             getInstance(start).addState(CREATING_NEW_TABLE, null);
             update(con, PasswordlessQueries.getQueryToCreateUsersTable(start), NO_OP_SETTER);
+
+            // index
+            update(con, PasswordlessQueries.getQueryToCreatePasswordlessUsersEmailIndex(start), NO_OP_SETTER);
+            update(con, PasswordlessQueries.getQueryToCreatePasswordlessUsersPhoneNumberIndex(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getPasswordlessUserToTenantTable())) {
             getInstance(start).addState(CREATING_NEW_TABLE, null);
             update(con, PasswordlessQueries.getQueryToCreatePasswordlessUserToTenantTable(start),
                     NO_OP_SETTER);
+
+            // index
+            update(con, PasswordlessQueries.getQueryToCreatePasswordlessUserToTenantEmailIndex(start), NO_OP_SETTER);
+            update(con, PasswordlessQueries.getQueryToCreatePasswordlessUserToTenantPhoneNumberIndex(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getPasswordlessDevicesTable())) {
@@ -419,6 +437,15 @@ public class GeneralQueries {
             update(con, TOTPQueries.getQueryToCreateUsedCodesTable(start), NO_OP_SETTER);
             // index:
             update(con, TOTPQueries.getQueryToCreateUsedCodesExpiryTimeIndex(start), NO_OP_SETTER);
+        }
+
+        if (!doesTableExists(start, con, Config.getConfig(start).getBulkImportUsersTable())) {
+            getInstance(start).addState(CREATING_NEW_TABLE, null);
+            update(start, BulkImportQueries.getQueryToCreateBulkImportUsersTable(start), NO_OP_SETTER);
+            // index:
+            update(start, BulkImportQueries.getQueryToCreateStatusUpdatedAtIndex(start), NO_OP_SETTER);
+            update(start, BulkImportQueries.getQueryToCreatePaginationIndex1(start), NO_OP_SETTER);
+            update(start, BulkImportQueries.getQueryToCreatePaginationIndex2(start), NO_OP_SETTER);
         }
 
         if (!doesTableExists(start, con, Config.getConfig(start).getOAuthClientsTable())) {
@@ -720,7 +747,7 @@ public class GeneralQueries {
                         // attach email tags to queries
                         QUERY = QUERY +
                                 " WHERE (emailpasswordTable.app_id = ? AND emailpasswordTable.tenant_id = ?) AND"
-                                + " (emailpasswordTable.email LIKE ? OR emailpasswordTable.email LIKE ?)";
+                                + " (emailpasswordTable.email LIKE ? OR emailpasswordTable.email LIKE ? ";
                         queryList.add(tenantIdentifier.getAppId());
                         queryList.add(tenantIdentifier.getTenantId());
                         queryList.add(dashboardSearchTags.emails.get(0) + "%");
@@ -730,6 +757,8 @@ public class GeneralQueries {
                             queryList.add(dashboardSearchTags.emails.get(i) + "%");
                             queryList.add("%@" + dashboardSearchTags.emails.get(i) + "%");
                         }
+
+                        QUERY += " )";
 
                         USER_SEARCH_TAG_CONDITION.append("SELECT * FROM ( ").append(QUERY)
                                 .append(" LIMIT 1000) AS emailpasswordResultTable");
@@ -1029,6 +1058,32 @@ public class GeneralQueries {
         }
     }
 
+    public static void makePrimaryUsers_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
+                                                    List<String> userIds)
+            throws SQLException, StorageQueryException {
+
+        String users_update_QUERY = "UPDATE " + Config.getConfig(start).getUsersTable() +
+                " SET is_linked_or_is_a_primary_user = true WHERE app_id = ? AND user_id = ?";
+        String appid_to_userid_update_QUERY = "UPDATE " + Config.getConfig(start).getAppIdToUserIdTable() +
+                " SET is_linked_or_is_a_primary_user = true WHERE app_id = ? AND user_id = ?";
+
+        List<PreparedStatementValueSetter> usersUpdateBatch = new ArrayList<>();
+        List<PreparedStatementValueSetter> appIdToUserIdUpdateBatch = new ArrayList<>();
+
+        for(String userId: userIds){
+            usersUpdateBatch.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, userId);
+            });
+            appIdToUserIdUpdateBatch.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, userId);
+            });
+        }
+        executeBatch(sqlCon, users_update_QUERY, usersUpdateBatch);
+        executeBatch(sqlCon, appid_to_userid_update_QUERY, appIdToUserIdUpdateBatch);
+    }
+
     public static void linkAccounts_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
                                                 String recipeUserId, String primaryUserId)
             throws SQLException, StorageQueryException {
@@ -1057,6 +1112,69 @@ public class GeneralQueries {
                 pst.setString(3, recipeUserId);
             });
         }
+    }
+
+    public static void linkMultipleAccounts_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
+                                                        Map<String, String> recipeUserIdToPrimaryUserId)
+            throws SQLException, StorageQueryException {
+
+        if(recipeUserIdToPrimaryUserId == null || recipeUserIdToPrimaryUserId.isEmpty()){
+            return;
+        }
+
+        String update_users_QUERY = "UPDATE " + Config.getConfig(start).getUsersTable() +
+                " SET is_linked_or_is_a_primary_user = true, primary_or_recipe_user_id = ? WHERE app_id = ? AND " +
+                "user_id = ?";
+
+        String update_appid_to_userid_QUERY = "UPDATE " + Config.getConfig(start).getAppIdToUserIdTable() +
+                " SET is_linked_or_is_a_primary_user = true, primary_or_recipe_user_id = ? WHERE app_id = ? AND " +
+                "user_id = ?";
+
+        List<PreparedStatementValueSetter> usersUpdateBatch = new ArrayList<>();
+        List<PreparedStatementValueSetter> appIdToUserIdUpdateBatch = new ArrayList<>();
+
+        for(Map.Entry<String, String> linkEntry : recipeUserIdToPrimaryUserId.entrySet()) {
+            String primaryUserId = linkEntry.getValue();
+            String recipeUserId = linkEntry.getKey();
+
+            usersUpdateBatch.add(pst -> {
+                pst.setString(1, primaryUserId);
+                pst.setString(2, appIdentifier.getAppId());
+                pst.setString(3, recipeUserId);
+            });
+            appIdToUserIdUpdateBatch.add(pst -> {
+                pst.setString(1, primaryUserId);
+                pst.setString(2, appIdentifier.getAppId());
+                pst.setString(3, recipeUserId);
+            });
+        }
+
+        executeBatch(sqlCon, update_users_QUERY, usersUpdateBatch);
+        executeBatch(sqlCon, update_appid_to_userid_QUERY, appIdToUserIdUpdateBatch);
+
+        updateTimeJoinedForPrimaryUsers_Transaction(start, sqlCon, appIdentifier,
+                new ArrayList<>(recipeUserIdToPrimaryUserId.values()));
+    }
+
+    public static void updateTimeJoinedForPrimaryUsers_Transaction(Start start, Connection sqlCon,
+                                                                   AppIdentifier appIdentifier, List<String> primaryUserIds)
+            throws SQLException, StorageQueryException {
+
+        String QUERY = "UPDATE " + Config.getConfig(start).getUsersTable() +
+                " JOIN (SELECT primary_or_recipe_user_id, MIN(time_joined) AS min_time_joined FROM " +
+                Config.getConfig(start).getUsersTable() + " WHERE app_id = ? AND primary_or_recipe_user_id = ? )" +
+                " AS temp_table ON " + Config.getConfig(start).getUsersTable() + ".primary_or_recipe_user_id = " +
+                "temp_table.primary_or_recipe_user_id " +
+                " SET primary_or_recipe_user_time_joined = min_time_joined";
+
+        List<PreparedStatementValueSetter> updateBatch = new ArrayList<>();
+        for(String primaryUserId : primaryUserIds) {
+            updateBatch.add(pst -> {
+                pst.setString(1, appIdentifier.getAppId());
+                pst.setString(2, primaryUserId);
+            });
+        }
+        executeBatch(sqlCon, QUERY, updateBatch);
     }
 
     public static void unlinkAccounts_Transaction(Start start, Connection sqlCon, AppIdentifier appIdentifier,
@@ -1266,17 +1384,29 @@ public class GeneralQueries {
                                                                           TenantIdentifier tenantIdentifier,
                                                                           String credentialId)
             throws StorageQueryException, SQLException, StorageTransactionLogicException {
-        // TODO: revisit this. Seems like we are loading the same data multiple times
         AuthRecipeUserInfo webauthnUser = start.startTransaction(con -> {
             try {
                 Connection sqlCon = (Connection) con.getConnection();
-                return WebAuthNQueries.getUserInfoByCredentialId_Transaction(start, sqlCon, tenantIdentifier,
+                return getPrimaryUserByWebauthNCredentialId_Transaction(start, sqlCon, tenantIdentifier,
                         credentialId);
             } catch (SQLException e) {
                 throw new StorageQueryException(e);
             }
         });
         return getPrimaryUserInfoForUserId(start, tenantIdentifier.toAppIdentifier(), webauthnUser.getSupertokensUserId());
+    }
+
+    public static AuthRecipeUserInfo getPrimaryUserByWebauthNCredentialId_Transaction(Start start,
+                                                                          Connection connection,
+                                                                          TenantIdentifier tenantIdentifier,
+                                                                          String credentialId)
+            throws StorageQueryException, SQLException, StorageTransactionLogicException {
+
+        AuthRecipeUserInfo webauthnUser = WebAuthNQueries.getUserInfoByCredentialId_Transaction(start, connection,
+                    tenantIdentifier, credentialId);
+
+        return getPrimaryUserInfoForUserId_Transaction(start, connection, tenantIdentifier.toAppIdentifier(),
+                webauthnUser.getSupertokensUserId());
     }
 
     public static String getPrimaryUserIdStrForUserId(Start start, AppIdentifier appIdentifier, String id)
@@ -1413,6 +1543,17 @@ public class GeneralQueries {
 
         return userIdToAuthRecipeUserInfo.keySet().stream().map(userIdToAuthRecipeUserInfo::get)
                 .collect(Collectors.toList());
+    }
+
+    public static List<AuthRecipeUserInfo> getPrimaryUserInfosForUserIds_Transaction(Start start, Connection con,
+                                                                                     AppIdentifier appIdentifier, List<String> ids)
+            throws SQLException, StorageQueryException {
+
+        List<AuthRecipeUserInfo> result = getPrimaryUserInfoForUserIds_Transaction(start, con, appIdentifier, ids);
+        if (result.isEmpty()) {
+            return null;
+        }
+        return result;
     }
 
     private static List<AuthRecipeUserInfo> getPrimaryUserInfoForUserIds_Transaction(Start start, Connection sqlCon,
@@ -1748,6 +1889,24 @@ public class GeneralQueries {
             pst.setString(1, appIdentifier.getAppId());
         }, result -> {
             return result.next();
+        });
+    }
+
+    public static List<String> findUserIdsThatExist(Start start, AppIdentifier appIdentifier, List<String> userIds)
+            throws SQLException, StorageQueryException {
+        String QUERY = "SELECT user_id FROM " + Config.getConfig(start).getAppIdToUserIdTable()
+                + " WHERE app_id = ? AND user_id IN ("+ Utils.generateCommaSeperatedQuestionMarks(userIds.size()) +")";
+        return execute(start, QUERY, pst -> {
+            pst.setString(1, appIdentifier.getAppId());
+            for(int i = 0; i<userIds.size(); i++) {
+                pst.setString(2 + i, userIds.get(i));
+            }
+        }, result -> {
+            List<String> foundUserIds = new ArrayList<>();
+            while(result.next()){
+                foundUserIds.add(result.getString(1));
+            }
+            return foundUserIds;
         });
     }